+++ test/browser/components/kepler-gl-test.js
@@ -42,7 +42,6 @@ import NotificationPanelFactory from 'components/notification-panel';
 import {ActionTypes} from 'actions';
 import {DEFAULT_MAP_STYLES, EXPORT_IMAGE_ID} from 'constants';
 import {GEOCODER_DATASET_NAME} from 'constants/default-settings';
-import {filterOutGeocoderDataset, sidePanelSelector} from 'components/kepler-gl';
 // mock state
 import {StateWithGeocoderDataset} from 'test/helpers/mock-state';
 
@@ -480,53 +479,6 @@ test('Components -> KeplerGl -> Mount -> Load custom map style task', t => {
   t.end();
 });
 
-// TODO: nikola move this into node tests utils
-test('Components -> KeplerGl -> Helpers -> Filter out geocoder dataset', t => {
-  // Geocoder dataset mock can be an empty object since the filter function only cares about the key
-  // in the 'datasets' object and filters by it
-  const datasets = {
-    geocoder_dataset: {},
-    first: {},
-    second: {}
-  };
-
-  t.true(
-    datasets[GEOCODER_DATASET_NAME],
-    `${GEOCODER_DATASET_NAME} key should exist before being filtered`
-  );
-
-  const filteredResults = filterOutGeocoderDataset(datasets);
-
-  t.isEqual(
-    filteredResults[GEOCODER_DATASET_NAME],
-    undefined,
-    `Should not exist after filtering out ${GEOCODER_DATASET_NAME} key`
-  );
-
-  t.end();
-});
-
-// TODO: nikola move this into node tests utils
-test('Components -> KeplerGl -> sidePanelSelector -> Do not pass geocoder dataset to SidePanel', t => {
-  // We don't need available providers sinde we are not referencing them
-  const sideFields = sidePanelSelector(StateWithGeocoderDataset, undefined);
-
-  t.isEqual(
-    sideFields.datasets[GEOCODER_DATASET_NAME],
-    undefined,
-    `SidePanelSelector should filter out ${GEOCODER_DATASET_NAME} from datasets object`
-  );
-
-  // tooltip
-  t.isEqual(
-    sideFields.interactionConfig.tooltip.config.fieldsToShow[GEOCODER_DATASET_NAME],
-    undefined,
-    `SidePanelSelector should filter out ${GEOCODER_DATASET_NAME} from tooltip fields to be shown`
-  );
-
-  t.end();
-});
-
 // TODO
 // Test data has only the 'geocoder_dataset' dataset
 // This function will return its name if it finds the dataset
@@ -547,7 +499,7 @@ test('Components -> KeplerGl -> SidePanel -> Geocoder dataset display', t => {
 
   const toggleSidePanel = sinon.spy();
 
-  // Create custom SidePanel that will get PanelToggle injected and which will be injected in custom KeplerGl component
+  // Create custom SidePanel that will accept toggleSidePanel as a spy
   function CustomSidePanelFactory(...deps) {
     const OriginalSidePanel = SidePanelFactory(...deps);
     const CustomSidePanel = props => {
